Make it easier to write new scheduling scripts
This separates out some common infrastructure from the pktgen scheduling script
